[build] Exclude some e2e suites from experiments tests
We have a bunch of experiments testing against core e2e - and this is
good for robustness, bad for CI times.

We also have a bunch of marginal but overall necessary fixtures in the
e2e suites - again good for robustness, bad for CI times.

We can eliminate some of the cross product though, and I think safely:
run experiments on a broad range of suites, but not *ALL* the suites,
and get a bunch of our CI time back.

Here I introduce an environment variable: `GRPC_CI_EXPERIMENTS` that's
set when running bazel @experiment= configs, cleared otherwise (so we
can still execute those tests directly when necessary). When that env
var is set we filter out a bunch of suites from the test configurations.