[chttp2] Dont send ping-acks with large data frames (#34583)
If we send a large amount of data along with an ack, then we bundle that
ack in with a crypto frame that might be very large. This causes two
problems:
1. we need to unencrypt a large frame (up to 1MB with ALTS) before we
can ack the ping, and on a slow connection this could take some time.
2. we need to do all the crypto work up front to send that ping ack,
also creating lots of work.

I think there's an equivalent fix needed on the ping send side, but less
urgent because it's not currently causing flakiness!