GrpcPolledFdWindowsFactory owns GrpcPolledFdWindows
AresResolver has a raw pointer of GrpcPolledFdWindows.

GrpcPolledFdWindowsFactory never destroys GrpcPolledFdWindows except
when it is destroyed.

The socket is closed in the destructor or when there is an external
cancelation.

The entry in the socket map does not need to be cleared in the latter
case since there won't be any new request.