[chttp2] Alternative protection for too many streams in the system (#34697)
Cancel streams if we have too much concurrency on a single channel to
allow that server to recover.

There seems to be a convergence in the HTTP2 community about this being
a reasonable thing to do, so I'd like to try it in some real scenarios.

If this pans out well then I'll likely drop the
`red_max_concurrent_streams` and the `rstpit` experiments in preference
to this.

I'm also considering tying in resource quota so that under high memory
pressure we just default to this path.

---------

Co-authored-by: ctiller <ctiller@users.noreply.github.com>