Fix breakage when built against protobuf head
A recent change to protobuf's `rules_python` version caused gRPC's Bazel
build to break. Part of the problem seemed to be that gRPC used a
different name for the dependency, `io_bazel_rules_python`, and this
resulted in two copies of `rules_python` getting included in the build.
This commit makes gRPC use the `rules_python` name so that it can choose
its own version to pin and so that there's only one copy. That resulted
in another breakage, but I was able to fix that by upgrading to the most
recent release. I dropped the patch to `rules_python`, since hopefully
it is no longer needed.

Another necessary fix was to update references to `@utf8_range`, since
this has been merged into the protobuf Bazel repo.