[upb] cast to (upb_Message*) when calling upb functions
upb currently defines upb_Message to be void, which has allowed us to be lax
with the types when calling upb functions that take a (upb_Message*) as an arg.

If this ever changes (which may happen quite soon) things will break, so this
is a pre-emptive patch to protect against such a change.