[tls] Fix incorrect use of certificate verification callbacks
As documented in [0], there are two certificate verification callbacks
in the OpenSSL/BoringSSL TLS API. The one taken as a parameter to
SSL_CTX_set_verify is the "verify callback". It is called multiple times
during a single certificate verification is used to suppress errors and
otherwise be notified about various events during verification.

Such a callback is not appropriate for accepting all certificates (you
waste time processing things that will be thrown away), nor for
post-verification inspection of the result (it will run multiple times).
This is, however, what gRPC does with it.

Rather, gRPC should have used SSL_CTX_set_cert_verify_callback, which
swaps out the verification process entirely. That is called exactly once
per handshake and allows you to skip the verification, or verify and
then inspect the results afterwards. Fix gRPC to heed the documentation.

In addition, this PR fixes a lifetime bug in gRPC's handling of the root
certificate. RootCertExtractCallback stashes the root certificate
without retaining it anywhere, but the X509_STORE_CTX will shortly be
destroyed. There is no immediate guarantee the X509 object lasts as long
as the SSL object. It most likely does because the object is often
cached in the X509_STORE, which lives on the SSL_CTX, but this is at
best, non-obvious. Instead, gRPC should have made
g_ssl_ex_verified_root_cert_index own a refcount to the X509 object by
registering a free function and calling X509_up_ref when saving the
value.

[0] https://commondatastorage.googleapis.com/chromium-boringssl-docs/ssl.h.html#SSL_CTX_set_verify