[promises] Add a switch primitive
Remove the old `switch` library - this used to be an implementation
detail of `Seq`, `TrySeq` - but has become unused.

Add a new user facing primitive `Switch` that fills a similar role to
`switch` in C++ - selecting a promise to execute based on a primitive
discriminator - much like `If` allows selection based on a boolean
discriminator now.

A future change will optimize this to actually lower the `Switch` into
an actual `switch` statement, but for right now I want to get the
functionality in.