[unix path] fix scheme slashes + absolute path with systemd activation
According to doc/naming.md, when a unix path can be
- `unix:relative`, example `unix:foo` or `unix:../bar`
- `unix:absolute`, example `unix:/tmp/foo`
- `unix://absolute`, example `unix:///tmp/foo`

Each form works when not using systemd socket activation.
But with systemd socket activation, the `unix:///tmp/foo`
is not working, because `sd_is_socket_unix` does not seem
to handle the remaining `///tmp/foo` and does not identify
it as matching one of the provided file descriptor.
This fixes that case

I included an additional error management for the case
where only two slashes are provided, e.g unix://tmp/foo,
which as per doc/naming.md, should be invalid in any case.

release notes: no