[EventEngine] rm invalid assert (#35953) (#36067)
Backports #35953 to v1.62.x

----

In the case where the timer delay is <= 0, RunAfter will return an
invalid handle since the callback will not be cancellable. The asserting
is invalid in this case. Based on how this handle is used, it seems
there is no need for a check that the handle is set. It would only
matter if `finish_bdp_ping` were called again before
`next_bdp_ping_timer_expired` were called, which I don't think is
possible.