[export] [security] Reland: Refactor credentials types to remove special handling for insecure creds
This PR refactors the credentials types to remove Secure and Insecure Channel and Call credentials types. We standardize on a `c_creds()` accessor method for all credentials types, which can now be treated uniformly. This notably removes special-case handling of insecure credentials.

The special code-paths for insecure creds are no longer necessary after #25586.

----
DO NOT SUBMIT. This PR is for testing purposes only. [cl/621192395](http://cl/621192395)

PiperOrigin-RevId: 621192395