Fix fuzzer builds internally for core e2e tests
The build layout I chose in https://github.com/grpc/grpc/pull/36198 causes problems for Google's internal build systems. This change fixes things so that link order is correct - unfortunately at the cost of adding 90-some C++ compilations.

I've made a new .cc file that is as minimal as possible to reduce the cost of those compilations down as far as I can.

PiperOrigin-RevId: 623012695