[context] Forcibly inline fast paths (#36970)
When building #36946 I noticed a bunch of places where `GetContext<>` was not inlined. This change forcibly inlines paths that we need to be, and additionally lowers the `CHECK` in `GetContext` to be a `DCHECK` as this comparison was using the bulk of the time in a very hot function in the stack.

Closes #36970

COPYBARA_INTEGRATE_REVIEW=https://github.com/grpc/grpc/pull/36970 from ctiller:dcheckctx 28a779899db678c23196d7753877701c22d77faf
PiperOrigin-RevId: 645202335