[Gpr_To_Absl_Logging] Cleaning up the allow list for banned functions (#36977)
[Gpr_To_Absl_Logging] Cleaning up the allow list for deprecated functions.

This file has a list of deprecated functions.

Any new instance of a deprecated function being used in the code will be flagged by the script. If there is a new instance of a deprecated function in a Pull Request, then the Sanity tests will fail.

We are currently working on clearing out the usage of deprecated functions in our code base.

For that reason, while our cleaning is in progress we have a temporary allow list. The allow list has a list of files where clean up of deprecated functions is pending. As we clean up the deprecated function from files , we will remove them from the allow list.

Ideally, it would be wise to do the file clean up and the altering of the allow list int the same PR. This will make sure that any roll back of a clean up PR will alter the allow list and avoid build failures.

Closes #36977

COPYBARA_INTEGRATE_REVIEW=https://github.com/grpc/grpc/pull/36977 from tanvi-jagtap:clean_allow_list_2024_06_19 9393244cee540977adb95b4b4120fdd85e0f408b
PiperOrigin-RevId: 645325296